[cleanup] erefactor/EclipseJdt - Use Autoboxing
EclipseJdt cleanup 'UseAutoboxing' applied by erefactor.

For EclipseJdt see https://www.eclipse.org/eclipse/news/4.19/jdt.php
For erefactor see https://github.com/cal101/erefactor

+++ hutool-core/src/main/java/cn/hutool/core/lang/mutable/MutableShort.java
@@ -46,7 +46,7 @@ public class MutableShort extends Number implements Comparable<MutableShort>, Mu
 
 	@Override
 	public Short get() {
-		return Short.valueOf(this.value);
+		return this.value;
 	}
 
 	/**
